Minor settings changes + some deprecation solved
ArUcoKeypointsDetector now uses parent settings file;
SurfKeypointDetector now uses parent settings file, also some syntax
changes added; KeypointsRejection usage of parent settings file added,
some warnings and deprecations solved; OpenNI some IO memory errors
solved, usage of parent settings file added, some warnings solved;
SacRegistration now uses parent settings file, some deprecation errors
solved; IcpRegistration now uses parent settings file;
ReconstructionInterface changed according to above changes, also added
separate key-points rejection function
Note: untested;

+++ imagesviewerwidget.cpp
@@ -8,7 +8,7 @@ ImagesViewerWidget::ImagesViewerWidget(std::vector<cv::Mat> images_vector, QStri
 	{
 		title = window_title;
 
-		for (int i = 0; i < images_vector.size(); i++)
+		for (size_t i = 0; i < images_vector.size(); i++)
 			imagesVector.push_back(Mat2QImage(images_vector[i]));
 
 		image_index = 0;
